requirements.txt
python-dotenv
customtkinter
psutil
pyaudio
numpy
opencv-python
pyautogui
Pillow
pyscreeze
edge-tts
SpeechRecognition
google-genai
openai
pygame
fastapi
uvicorn
cryptography
requests
pyinstaller
